requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-sqlalchemy==3.0.3
bcrypt==4.0.1
python-dotenv==1.0.0
pandas==2.0.0
openpyxl==3.1.2
xlrd==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
requests==2.28.2
email-validator==2.0.0
Pillow==9.5.0
